refactor(dataviews): keep each core domain's types in its types.ts
Seventy-one types were declared beside the factory that first needed
them, nineteen of them imported from other files by the factory's path.
Every domain of the core now holds its exported types in one `types.ts`,
with a file-level block saying what binds them, and an implementation
file exports its one function; the domain barrels re-export the types
file whole and each function by name. The structural IndexedDB surface
the store drives moves with the factory type that names it.
