+++ packages/runtime/dataviews/src/lib/collection/index.ts
@@ -1,9 +1,2 @@
-export type {
-  CollectionCoordinator,
-  CollectionCoordinatorConfig,
-  CollectionState,
-  DispatchResult,
-  ResultState,
-  ResultStatus,
-} from "./createCollectionCoordinator.js";
 export { default as createCollectionCoordinator } from "./createCollectionCoordinator.js";
+export type * from "./types.js";
