+++ packages/runtime/dataviews/src/lib/field/index.ts
@@ -1,8 +1,2 @@
-export type {
-  FieldFeedback,
-  FieldInteraction,
-  FieldInteractionConfig,
-  FieldInteractionState,
-  FieldValidation,
-} from "./createFieldInteraction.js";
 export { default as createFieldInteraction } from "./createFieldInteraction.js";
+export type * from "./types.js";
