+++ packages/runtime/dataviews/src/lib/indexeddb/index.ts
@@ -1,5 +1,5 @@
+export { default as createIndexedDBViewStore } from "./createIndexedDBViewStore.js";
 export type {
   IndexedDBFactory,
   IndexedDBViewStoreConfig,
-} from "./createIndexedDBViewStore.js";
-export { default as createIndexedDBViewStore } from "./createIndexedDBViewStore.js";
+} from "./types.js";
