+++ packages/runtime/dataviews/src/lib/observable/index.ts
@@ -1,6 +1,2 @@
-export type {
-  Channel,
-  ChannelConfig,
-  ReadonlyChannel,
-} from "./createChannel.js";
 export { default as createChannel } from "./createChannel.js";
+export type * from "./types.js";
