+++ packages/runtime/dataviews/src/lib/operation/index.ts
@@ -1,9 +1,2 @@
-export type {
-  ActionInvocation,
-  Operation,
-  OperationConfig,
-  OperationFailure,
-  OperationOutcome,
-  OperationState,
-} from "./createOperation.js";
 export { default as createOperation } from "./createOperation.js";
+export type * from "./types.js";
