+++ packages/runtime/dataviews/src/lib/selection/index.ts
@@ -1,2 +1,2 @@
-export type { Selection, SelectionState } from "./createSelection.js";
 export { default as createSelection } from "./createSelection.js";
+export type * from "./types.js";
